zelf_hackathon/settings/__init__.py
from pathlib import Path

import environ
import os
from .defaults import env

if env.str('ENV_TYPE') == 'STAGING':
    from .staging import *
elif env.str('ENV_TYPE') == 'PRODUCTION':
    from .production import *
else:
    from .development import *
